Got the reading and writing of a basic CSV file implemented.
Re-arranged the program slightly to have there be a series of 
subcommands based on which latch switch is activated.
You then choose the timer functionality from that, and also updated the 
"reset" functionality of the enter button to end the timer if active - 
save it out - and make the elements invisible again.